Integrated buffoon protobuf changes

+++ ext/extconf.rb
@@ -3,41 +3,47 @@ require 'rbconfig'
 require 'net/http'
 require 'zlib'
 
-unless RbConfig::CONFIG["arch"] == "x86_64-linux"
-  puts "At the moment, Skylight only supports 64-bit linux"
-  File.open("Makefile", "w") { |f| f.puts "hello:\ninstall:" }
+unless File.exist?("libskylight.a")
+  unless RbConfig::CONFIG["arch"] == "x86_64-linux"
+    puts "At the moment, Skylight only supports 64-bit linux"
+    File.open("Makefile", "w") { |f| f.puts "hello:\ninstall:" }
 
-  exit
+    exit
+  end
+
+  location = nil
+
+  Net::HTTP.start("github.com", 443, use_ssl: true) do |http|
+    uri = URI.parse("https://github.com/skylightio/skylight-rust/releases/download/v0.3.0-pre.1/libskylight-x86_64-linux.a.gz")
+    response = http.get(uri.request_uri)
+    location = response["Location"]
+  end
+
+  archive = nil
+
+  uri = URI(location)
+  Net::HTTP.start(uri.host, uri.port, use_ssl: true) do |http|
+    response = http.get(uri.request_uri)
+    archive = response.body
+  end
+
+  inflater, dest = Zlib::Inflate.new(32 + Zlib::MAX_WBITS), ""
+  dest << inflater.inflate(archive)
+  inflater.close
+
+  File.open("libskylight.a", "w") { |file| file.write dest }
 end
-
-location = nil
-
-Net::HTTP.start("github.com", 443, use_ssl: true) do |http|
-  uri = URI.parse("https://github.com/skylightio/skylight-rust/releases/download/v0.3.0-pre.1/libskylight-x86_64-linux.a.gz")
-  response = http.get(uri.request_uri)
-  location = response["Location"]
-end
-
-archive = nil
-
-uri = URI(location)
-Net::HTTP.start(uri.host, uri.port, use_ssl: true) do |http|
-  response = http.get(uri.request_uri)
-  archive = response.body
-end
-
-inflater, dest = Zlib::Inflate.new(32 + Zlib::MAX_WBITS), ""
-dest << inflater.inflate(archive)
-inflater.close
-
-File.open("libskylight.a", "w") { |file| file.write dest }
 
 have_header 'dlfcn.h'
 
 find_library("skylight", "factory", ".")
 
-$LDFLAGS << " -Wl,--version-script=skylight.map"
-$LDFLAGS << " -lrt -ldl -lm -lpthread"
+if Gem::Platform.local.os == "darwin"
+  $LDFLAGS << " -lpthread"
+else
+  $LDFLAGS << " -Wl,--version-script=skylight.map"
+  $LDFLAGS << " -lrt -ldl -lm -lpthread"
+end
 
 CONFIG['warnflags'].gsub!('-Wdeclaration-after-statement', '')
